requirements.txt
# Core
torch>=2.2.0
torchvision>=0.17.0
ultralytics>=8.2.0
timm>=1.0.0

# Data / augmentation
albumentations>=1.4.0
opencv-python>=4.9.0
numpy>=1.26.0
Pillow>=10.0.0
pycocotools>=2.0.7

# Schema / config
pydantic>=2.7.0
PyYAML>=6.0.1

# Dashboard
streamlit>=1.35.0

# Testing
pytest>=8.0.0

# Phase 3 (optional — install manually when needed)
# paddlepaddle
# paddleocr

# Phase 2 (optional)
# lapx>=0.5.5   # ByteTrack dependency
